Update to latest

+++ cloudflare/CloudflarePlugin.php
@@ -11,7 +11,7 @@ class CloudflarePlugin extends BasePlugin
 
 	public function getVersion()
 	{
-		return '0.1.1';
+		return '0.1.4';
 	}
 
 	public function getSchemaVersion()
@@ -21,7 +21,7 @@ class CloudflarePlugin extends BasePlugin
 
 	public function getDescription()
 	{
-		return 'Automatically purge Cloudflare URLs.';
+		return 'Purge Cloudflare caches from Craft.';
 	}
 
 	public function getDeveloper()
@@ -51,30 +51,36 @@ class CloudflarePlugin extends BasePlugin
 
 	public function init()
 	{
-		craft()->on('assets.onSaveAsset', function (Event $event) {
-			if ($event->params['isNewAsset'] === false)
-			{
+		if (craft()->cloudflare->settings->purgeAssetUrls)
+		{
+			craft()->on('assets.onSaveAsset', function (Event $event) {
+				if ($event->params['isNewAsset'] === false)
+				{
+					$asset = $event->params['asset'];
+					craft()->cloudflare->purgeUrls(array($asset->url));
+				}
+			});
+
+			craft()->on('assets.onDeleteAsset', function (Event $event) {
 				$asset = $event->params['asset'];
 				craft()->cloudflare->purgeUrls(array($asset->url));
-			}
-		});
+			});
+		}
 
-		craft()->on('assets.onDeleteAsset', function (Event $event) {
-			$asset = $event->params['asset'];
-			craft()->cloudflare->purgeUrls(array($asset->url));
-		});
+		if (craft()->cloudflare->settings->purgeEntryUrls)
+		{
+			craft()->on('entries.onSaveEntry', function (Event $event) {
+				$entry = $event->params['entry'];
+				craft()->cloudflare->purgeUrls(array($entry->url));
+				craft()->cloudflare_rules->purgeCachesForUrl($entry->url);
+			});
 
-		craft()->on('entries.onSaveEntry', function (Event $event) {
-			$entry = $event->params['entry'];
-			craft()->cloudflare->purgeUrls(array($entry->url));
-			craft()->cloudflare_rules->purgeCachesForUrl($entry->url);
-		});
-
-		craft()->on('entries.onDeleteEntry', function (Event $event) {
-			$entry = $event->params['entry'];
-			craft()->cloudflare->purgeUrls(array($entry->url));
-			craft()->cloudflare_rules->purgeCachesForUrl($entry->url);
-		});
+			craft()->on('entries.onDeleteEntry', function (Event $event) {
+				$entry = $event->params['entry'];
+				craft()->cloudflare->purgeUrls(array($entry->url));
+				craft()->cloudflare_rules->purgeCachesForUrl($entry->url);
+			});
+		}
 
 		// TODO: clear on publish, update, or delete entry
 	}
@@ -89,10 +95,38 @@ class CloudflarePlugin extends BasePlugin
 	protected function defineSettings()
 	{
 		return array(
-			'apiKey'         => array(AttributeType::String, 'required' => true, 'label' => 'Cloudflare API Key'),
-			'email'          => array(AttributeType::String, 'required' => true, 'label' => 'Cloudflare Account Email'),
-			'zone'           => array(AttributeType::String, 'required' => true, 'label' => 'Cloudflare Zone'),
-			'userServiceKey' => array(AttributeType::String, 'required' => false, 'label' => 'Cloudflare User Service Key'),
+			'apiKey' => array(
+				AttributeType::String,
+				'required' => true,
+				'label'    => 'Cloudflare API Key'
+			),
+			'email' => array(
+				AttributeType::String,
+				'required' => true,
+				'label'    => 'Cloudflare Account Email'
+			),
+			'zone' => array(
+				AttributeType::String,
+				'required' => true,
+				'label'    => 'Cloudflare Zone'
+			),
+			'purgeEntryUrls' => array(
+				AttributeType::Bool,
+				'required' => true,
+				'label'    => 'Purge Entry URLs',
+				'default'  => false
+			),
+			'purgeAssetUrls' => array(
+				AttributeType::Bool,
+				'required' => true,
+				'label'    => 'Purge Asset URLs',
+				'default'  => false
+			),
+			'userServiceKey' => array(
+				AttributeType::String,
+				'required' => false,
+				'label'    => 'Cloudflare User Service Key'
+			),
 		);
 	}
